backend/requirements.txt
requests
fastapi
uvicorn[standard]
python-dotenv
pymongo